+++ src/Commands/ManagementApi/GetOffice365Services.cs
@@ -15,7 +15,7 @@ namespace PnP.PowerShell.Commands.ManagementApi
     {
         protected override void ExecuteCmdlet()
         {
-            var collection = GraphHelper.GetAsync<GraphCollection<ManagementApiService>>(HttpClient, $"{ApiRootUrl}ServiceComms/Services", AccessToken, false).GetAwaiter().GetResult();
+            var collection = GraphHelper.GetAsync<RestResultCollection<ManagementApiService>>(HttpClient, $"{ApiRootUrl}ServiceComms/Services", AccessToken, false).GetAwaiter().GetResult();
 
             if(collection != null)
             {