+++ src/Commands/SiteDesigns/AddSiteScript.cs
@@ -1,6 +1,5 @@
 ﻿using Microsoft.Online.SharePoint.TenantAdministration;
 using Microsoft.SharePoint.Client;
-
 using PnP.PowerShell.Commands.Base;
 using System.Management.Automation;
 
@@ -13,7 +12,7 @@ namespace PnP.PowerShell.Commands
         public string Title;
 
         [Parameter(Mandatory = false)]
-        public string Description;     
+        public string Description;
 
         [Parameter(Mandatory = true)]
         public string Content;